updated oscpack to most recent version from http://code.google.com/p/oscpack/ (many fixes for x86_64)
git-svn-id: svn+ssh://svn.code.sf.net/p/tisch/code/libtisch@106 28952012-f250-0410-9090-e835bad21ca5

+++ libs/osc/osc/OscReceivedElements.cpp
@@ -491,7 +491,7 @@ void ReceivedMessageArgumentIterator::Advance()
         case BLOB_TYPE_TAG:
             {
                 uint32 blobSize = ToUInt32( value_.argument_ );
-                value_.argument_ = value_.argument_ + 4 + RoundUp4( blobSize );
+                value_.argument_ = value_.argument_ + 4 + RoundUp4( (unsigned long)blobSize );
             }
             break;
 
@@ -631,7 +631,7 @@ void ReceivedMessage::Init( const char *message, unsigned long size )
                                 MalformedMessageException( "arguments exceed message size" );
                                 
                             uint32 blobSize = ToUInt32( argument );
-                            argument = argument + 4 + RoundUp4( blobSize );
+                            argument = argument + 4 + RoundUp4( (unsigned long)blobSize );
                             if( argument > end )
                                 MalformedMessageException( "arguments exceed message size" );
                         }
